requirements.txt
tensorflow-gpu==2.11.0
scikit-learn==1.2.1
numpy==1.23.4
scipy==1.10.0
statsmodels==0.13.5